// OnlineExamSystem/src/app/Models/user.ts
export interface User {
            userId: number;
            name: string;
            emailId: string;
            mobile: string;
            state: string,
            city: string,
            dob: string;
            qualification: string;
            yearOfCompletion: string;
            password: string;
} 